requirements.txt
alembic==1.5.7
aniso8601==9.0.1
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
DAWG-Python==0.7.2
docopt==0.6.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
flask-ngrok==0.0.25
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
Pillow==8.2.0
pycparser==2.20
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dateutil==2.8.1
python-dotenv==0.16.0
python-editor==1.0.4
pytz==2021.1
qrcode==6.1
requests==2.25.1
six==1.15.0
speaklater==1.3
SQLAlchemy==1.4.1
SQLAlchemy-serializer==1.3.4.4
urllib3==1.26.4
waitress==2.0.0
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.3.0
